MSI-1440: There are no checks that prevent lost of the last association between Stock and Sales Channels when merchant tries to Save the Stock
