+++ app/code/Magento/InventorySalesApi/Test/Api/StockRepository/SalesChannelManagementTest.php
@@ -8,6 +8,7 @@ declare(strict_types=1);
 namespace Magento\InventorySalesApi\Test\Api\StockRepository;
 
 use Magento\Framework\Api\ExtensibleDataInterface;
+use Magento\Framework\Webapi\Exception;
 use Magento\Framework\Webapi\Rest\Request;
 use Magento\InventoryApi\Api\Data\StockInterface;
 use Magento\InventorySalesApi\Api\Data\SalesChannelInterface;
@@ -57,48 +58,181 @@ class SalesChannelManagementTest extends WebapiAbstract
 
     /**
      * @param array $salesChannels
+     * @param array $expectedErrorData
+     *
      * @magentoApiDataFixture ../../../../app/code/Magento/InventorySalesApi/Test/_files/websites_with_stores.php
      * @magentoApiDataFixture ../../../../app/code/Magento/InventorySalesApi/Test/_files/stock_with_sales_channels.php
-     * @dataProvider updateStockWithSalesChannelsReplacingDataProvider
+     *
+     * @dataProvider deleteSalesChannelDataProvider
      */
-    public function testUpdateStockWithSalesChannelsReplacing(array $salesChannels)
+    public function testValidateDeleteSalesChannelFromStock(array $salesChannels, array $expectedErrorData)
     {
         $stockId = 10;
-        $stockData = [
-            StockInterface::NAME => 'stock-name',
-            ExtensibleDataInterface::EXTENSION_ATTRIBUTES_KEY => [
-                'sales_channels' => $salesChannels,
+        $data = [
+            StockInterface::STOCK_ID => $stockId,
+            StockInterface::NAME => 'stock_with_channels_name',
+            'extension_attributes' => [
+                'sales_channels' => $salesChannels
+            ]
+        ];
+
+        $serviceInfo = [
+            'rest' => [
+                'resourcePath' => self::RESOURCE_PATH . '/' . $stockId,
+                'httpMethod' => Request::HTTP_METHOD_PUT,
+            ],
+            'soap' => [
+                'service' => self::SERVICE_NAME,
+                'operation' => self::SERVICE_NAME . 'Save',
             ],
         ];
-        $this->saveStock($stockId, $stockData);
-        $actualStockData = $this->getStockDataById($stockId);
 
-        self::assertArrayHasKey('sales_channels', $stockData[ExtensibleDataInterface::EXTENSION_ATTRIBUTES_KEY]);
-        self::assertEquals(
-            $salesChannels,
-            $actualStockData[ExtensibleDataInterface::EXTENSION_ATTRIBUTES_KEY]['sales_channels']
-        );
+        $this->webApiCallWithError($serviceInfo, $data, $expectedErrorData);
+    }
+
+    public function deleteSalesChannelDataProvider(): array
+    {
+        return [
+            'all_channels_delete' . SalesChannelInterface::TYPE => [
+                [
+                    [
+
+                        SalesChannelInterface::TYPE => SalesChannelInterface::TYPE_WEBSITE,
+                        SalesChannelInterface::CODE => 'us_website',
+
+                    ],
+                ],
+                [
+                    'message' => 'Validation Failed',
+                    'errors' => [
+                        [
+                            'message' => 'Website "%field" should be linked to stock.',
+                            'parameters' => [
+                                'field' => 'eu_website',
+                            ],
+                        ],
+                    ],
+                ],
+            ],
+        ];
+    }
+    
+    /**
+     * @param array $salesChannels
+     * @param array $expectedErrorData
+     *
+     * @magentoApiDataFixture ../../../../app/code/Magento/InventoryApi/Test/_files/stock.php
+     *
+     * @dataProvider dataProviderSalesChannelsAssignment
+     */
+    public function testFailedValidationSalesChannelsAssignment(array $salesChannels, array $expectedErrorData)
+    {
+        $stockId = 10;
+        $data = [
+            StockInterface::STOCK_ID => $stockId,
+            StockInterface::NAME => 'Default Stock',
+            'extension_attributes' => [
+                'sales_channels' => $salesChannels
+            ]
+        ];
+
+        $serviceInfo = [
+            'rest' => [
+                'resourcePath' => self::RESOURCE_PATH . '/' . $stockId,
+                'httpMethod' => Request::HTTP_METHOD_PUT,
+            ],
+            'soap' => [
+                'service' => self::SERVICE_NAME,
+                'operation' => self::SERVICE_NAME . 'Save',
+            ],
+        ];
+        $this->webApiCallWithError($serviceInfo, $data, $expectedErrorData);
     }
 
     /**
      * @return array
      */
-    public function updateStockWithSalesChannelsReplacingDataProvider(): array
+    public function dataProviderSalesChannelsAssignment(): array
     {
         return [
-            'replace_sales_channels' => [
+            'not_given_' . SalesChannelInterface::TYPE => [
                 [
                     [
-                        SalesChannelInterface::TYPE => SalesChannelInterface::TYPE_WEBSITE,
-                        SalesChannelInterface::CODE => 'global_website',
+                        SalesChannelInterface::CODE => 'base'
                     ],
-                    [
-                        SalesChannelInterface::TYPE => SalesChannelInterface::TYPE_WEBSITE,
-                        SalesChannelInterface::CODE => 'us_website',
+                ],
+                [
+                    'message' => 'Validation Failed',
+                    'errors' => [
+                        [
+                            'message' => '"%field" can not be empty.',
+                            'parameters' => [
+                                'field' => SalesChannelInterface::TYPE,
+                            ],
+                        ],
                     ],
                 ],
             ],
-            'remove_sales_channels' => [[]],
+            'empty_' . SalesChannelInterface::TYPE => [
+                [
+                    [
+                        SalesChannelInterface::TYPE => '',
+                        SalesChannelInterface::CODE => 'base'
+                    ],
+                ],
+                [
+                    'message' => 'Validation Failed',
+                    'errors' => [
+                        [
+                            'message' => '"%field" can not be empty.',
+                            'parameters' => [
+                                'field' => SalesChannelInterface::TYPE,
+                            ],
+                        ],
+                    ],
+                ],
+            ],
+            'not_given_' . SalesChannelInterface::CODE => [
+                [
+                    [
+                        SalesChannelInterface::TYPE => SalesChannelInterface::TYPE_WEBSITE,
+                        SalesChannelInterface::CODE => 'base'
+                    ],
+                    [
+                        SalesChannelInterface::TYPE => SalesChannelInterface::TYPE_WEBSITE,
+                    ],
+                ],
+                [
+                    'message' => 'Validation Failed',
+                    'errors' => [
+                        [
+                            'message' => '"%field" can not be empty.',
+                            'parameters' => [
+                                'field' => SalesChannelInterface::CODE,
+                            ],
+                        ],
+                    ],
+                ],
+            ],
+            'empty_' . SalesChannelInterface::CODE => [
+                [
+                    [
+                        SalesChannelInterface::TYPE => SalesChannelInterface::TYPE_WEBSITE,
+                        SalesChannelInterface::CODE => ''
+                    ],
+                ],
+                [
+                    'message' => 'Validation Failed',
+                    'errors' => [
+                        [
+                            'message' => '"%field" can not be empty.',
+                            'parameters' => [
+                                'field' => SalesChannelInterface::CODE,
+                            ],
+                        ],
+                    ],
+                ],
+            ],
         ];
     }
 
@@ -149,5 +283,38 @@ class SalesChannelManagementTest extends WebapiAbstract
             : $this->_webApiCall($serviceInfo, ['stockId' => $stockId]);
         self::assertArrayHasKey(StockInterface::STOCK_ID, $response);
         return $response;
+    }
+
+    /**
+     * @param array $serviceInfo
+     * @param array $data
+     * @param array $expectedErrorData
+     * @return void
+     * @throws \Exception
+     */
+    private function webApiCallWithError(array $serviceInfo, array $data, array $expectedErrorData)
+    {
+        try {
+            $this->_webApiCall($serviceInfo, ['stock' => $data]);
+            $this->fail('Expected throwing exception');
+        } catch (\Exception $e) {
+            if (TESTS_WEB_API_ADAPTER === self::ADAPTER_REST) {
+                self::assertEquals($expectedErrorData, $this->processRestExceptionResult($e));
+                self::assertEquals(Exception::HTTP_BAD_REQUEST, $e->getCode());
+            } elseif (TESTS_WEB_API_ADAPTER === self::ADAPTER_SOAP) {
+                $this->assertInstanceOf('SoapFault', $e);
+                $expectedWrappedErrors = [];
+                foreach ($expectedErrorData['errors'] as $error) {
+                    // @see \Magento\TestFramework\TestCase\WebapiAbstract::getActualWrappedErrors()
+                    $expectedWrappedErrors[] = [
+                        'message' => $error['message'],
+                        'params' => $error['parameters'],
+                    ];
+                }
+                $this->checkSoapFault($e, $expectedErrorData['message'], 'env:Sender', [], $expectedWrappedErrors);
+            } else {
+                throw $e;
+            }
+        }
     }
 }
